+++ internal/register/useremail.go
@@ -3,7 +3,7 @@ package register
 import "time"
 
 type UserEmail struct {
-	UserEmailId      int       `json:user_email_id`
+	UserEmailId      int64     `json:user_email_id`
 	UserId           int64     `json:user_id`
 	Email            string    `json:email`
 	NormalizedEmail  string    `json:normalized_email`
